==> 99_Results.ipynb <==
import numpy as np
import pandas as pd
from functions import *
from functions2 import *

d2 = pd.read_csv("results/constrained.csv")
i1 = d2['period'] == 'out-of-sample'
i2 = ( d2['portfolio'] == 'LS' ) | ( d2['portfolio'] == '5' )
i = i1 & i2
d2 = d2[i]

d0 = pd.read_csv("results/unconstrained.csv")
i1 = d0['period'] == 'out-of-sample'
i2 = ( d0['portfolio'] == 'LS' ) | ( d0['portfolio'] == '5' )
i = i1 & i2
d0 = d0[i]

d1 = pd.read_csv("results/lasso.csv")
i1 = d1['period'] == 'out-of-sample'
i2 = ( d1['portfolio'] == 'LS' ) | ( d1['portfolio'] == '5' )
i3 = d1['predictors'] <= 10
i = i1 & i2 & i3
d1 = d1[i].iloc[[-1],:]

d3 = pd.read_csv("results/model2_nonlinear_3d_mb.csv")
i1 = d3['period'] == 'out-of-sample'
i2 = ( d3['portfolio'] == 'LS' ) | ( d3['portfolio'] == '5' )
i = i1 & i2
if i.sum() > 0:
    d3 = d3[i]
d3

d4 = pd.read_csv("results/model3b_linear_IR_minibatches.csv")
i1 = d4['period'] == 'out-of-sample'
i2 = ( d4['portfolio'] == 'LS' ) | ( d4['portfolio'] == '5' )
i = i1 & i2
d4 = d4[i]

d5 = pd.read_csv("results/model6_nonlinear_IR.csv")
i1 = d5['period'] == 'out-of-sample'
i2 = ( d5['portfolio'] == 'LS' ) | ( d5['portfolio'] == '5' )
i = i1 & i2
d5 = d5[i]

d6 = pd.read_csv("results/mlfactor_perf.csv")
d6['model']     = 'xgboost'
d6['period']    = 'out-of-sample'
d6['portfolio'] = 'L'   # That is along-only portfolio -- it is not really comparable

d = pd.concat([d0,d1,d2,d3,d6,d4,d5])
tmp = d[['model','portfolio', 'period', 'CAGR', 'Annualized Volatility', 'Information Ratio', 'Turnover']].copy()
tmp.drop('period', axis=1, inplace=True)  # The table is too wide...
for i in ['CAGR', 'Annualized Volatility', 'Turnover']:
    tmp[i] = 100 * tmp[i]
tmp.columns = ['Model', 'Portfolio', 'CAGR (%)', 'AnnVol (%)', 'IR', 'Turnover (%)']

latex(
    tmp[ tmp['Portfolio'] == 'LS' ],
    align='llrrrr',
    #format = { u: latex_decimal if type( tmp[u].iloc[0] ) == np.float64  else str for u in tmp.columns },
    format = { 
        'CAGR (%)':     lambda u: latex_decimal(u,1),
        'AnnVol (%)':   lambda u: latex_decimal(u,1),
        'IR':           lambda u: latex_decimal(u,2),
        'Turnover (%)': lambda u: latex_decimal(u,0),
    },
    file = 'results/performance_all.tex'
)

latex(
    tmp[ tmp['Portfolio'] != 'LS' ],
    align='llrrrr',
    format = { 
        'CAGR (%)':     lambda u: latex_decimal(u,1),
        'AnnVol (%)':   lambda u: latex_decimal(u,1),
        'IR':           lambda u: latex_decimal(u,2),
        'Turnover (%)': lambda u: latex_decimal(u,0),
    },
    file = 'results/performance_all_long-only.tex'
)

display( tmp[ tmp['Portfolio'] != '5' ] )
display( tmp[ tmp['Portfolio'] != 'LS' ] )
